requirements.txt
pdf2image==1.17.0
pillow==11.3.0
numpy==2.2.6
img2pdf==0.6.1